+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 * **Login**: Adds option to disable username/password logins, closes [#4674](https://github.com/grafana/grafana/issues/4674)
+* **SingleStat**: Add seriename as option in singlestat panel, closes [#4740](https://github.com/grafana/grafana/issues/4740)
 
 
 # 3.1.0 stable (unreleased)